bugbot-jira/pyproject.toml
[tool.poetry]
name = "bugbot-jira"
version = "0.1.0"
description = "LLM-powered JIRA ticket generator"
authors = ["Marcis Abols <marcis.abols@gmail.com>"]
packages = [{ include = "bugbot", from = "src" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = { extras = ["standard"], version = "^0.29" }
pydantic = "^2.7"
jinja2 = "^3.1"
openai = "^1.30"
anthropic = ">=0.28,<0.29"
google-cloud-aiplatform = "^1.47"
opencv-python-headless = "^4.10"
pybase64 = "^1.3"
nltk = "^3.9"
atlassian-python-api = "^3.41.11"
python-dotenv = "^1.0"
nats-py = "^2.6"
psycopg = { version = "^3.2", extras = ["binary", "pool"] }
pgvector = "^0.2"
emoji = "^2.10"
python-multipart = "^0.0.9"
pydantic-settings = "^2.2"
httpx = "0.24.1"
google-generativeai = "^0.4"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2"
pytest-asyncio = "^0.23"
black = "^24.4"
ruff = "^0.4"
mypy = "^1.10"
commitizen = "^3.27"

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]
